Atk4 phpunit polyfill for phpunit < 7.0
