+++ src/AtkPhpunit/ResultPrinter.php
@@ -1,21 +1,26 @@
 <?php
 
-namespace atk4\core;
+namespace atk4\core\AtkPhpunit;
+
+use PHPUnit\Framework\TestFailure;
+use PHPUnit\TextUI\ResultPrinter;
+
+require_once __DIR__ . '/phpunit6_polyfill.php';
 
 /**
  * Generic ResultPrinter for PHPUnit tests of ATK4 repos.
  */
-class PHPUnit_AgileResultPrinter extends \PHPUnit_TextUI_ResultPrinter
+class ResultPrinter extends ResultPrinter
 {
     /**
      * Prints trace info.
      *
      * @param \PHPUnit_Framework_TestFailure $defect
      */
-    protected function printDefectTrace(\PHPUnit_Framework_TestFailure $defect): void
+    protected function printDefectTrace(TestFailure $defect): void
     {
         $e = $defect->thrownException();
-        if (!$e instanceof \atk4\core\PHPUnit_AgileExceptionWrapper) {
+        if (!$e instanceof ExceptionWrapper) {
             parent::printDefectTrace($defect);
 
             return;
